requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
deprecation==2.1.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pillow==10.2.0
postgrest==0.16.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==11.0.3
